Refactor WDIMatter section with styles file
